refactor original render
