+++ keyboards/lily58/keymaps/default/rules.mk
@@ -23,7 +23,8 @@ SLEEP_LED_ENABLE = no    # Breathing sleep LED during USB suspend
 SRC +=  ./lib/glcdfont.c \
         ./lib/layer_state_reader.c \
         ./lib/logo_reader.c \
-        ./lib/counter_render.c \
+        ./lib/uptime_render.c \
+        ./lib/typecount_render.c \
         ./lib/keylogger.c \
         # ./lib/mode_icon_reader.c \
         # ./lib/host_led_state_reader.c \
